Add a Jekyll site
This is not particularly optimal or easy to understand. I'm first going
to outline the way that this works (what's different vs a standard
Jekyll and GitHub Pages site), then I'm going to discuss some of the
reasoning behind my decisions.

How this works
--------------

Unlike a standard Jekyll/GitHub pages site, the source for the site
lives in a `site` directory on the `master` branch. This site is built
on CircleCI whenever a tag is pushed to the repository (or a GitHub
release creates a tag). as well as when something is pushed to the
`master` branch.

There are a few commands that are required to make this happen:

  - `make build-website-api`: this takes the current `CIRCLE_TAG`
    environment variable, infers a major version from it, and then saves
    the current competencies file at `/site/api/vX/competencies.json`
    (where `X` is the major version)

  - `make publish-website`: this does the following:
      1. Check out and pull the `gh-pages` branch
      2. Remove all tracked files in the repository to avoid conflicts
      3. Reset and check out important files:
          - `.circleci`: we don't want this to be overwritten or the
            builds will start failing on the `gh-pages` branch
          - `.gitignore`: this is required to make sure we don't commit
            any untracked files that may be present from the `master`
            branch, e.g. `node_modules`
          - `api`: the current `CIRCLE_TAG` may be a greater version
            than the existing API versions. If we remove this folder
            then old versions of the API would just disappear
      4. Copy the generated site (from `make build-website`) into the
         working directory. This means that the latest build of the
         website replaces everything except the excluded files mentioned
         above
      5. Commit all of the changes and push to the `gh-pages` branch
         on GitHub
      6. Check out the previous branch

Some justification
------------------

The way this works means that the following can happen:

  1. There is no manual process to copy the generated competencies JSON
     into the correct API endpoint
  2. Previous versions of the API aren't overwritten by new pushes to
     the site
  3. We can edit everything on one branch, no switching between
     `gh-pages` and `master`