Add YAML file for competencies and validate
This adds a YAML file for competencies and some of the tooling we need
around this. I've created a command to parse the YAML and save as JSON,
and a command to validate the generated JSON against the schema.

The following work needs to happen before this can be merged:

  - [ ] Get the proposal approved and reflect any changes in this code
  - [ ] Write technical documentation for the included scripts
  - [ ] Write placeholder documentation on how to edit competencies
        (the full documentation can be added after wordsmithing)
  - [ ] Add additional validation (if required) to make sure IDs are
        never duplicated and circular dependencies can't occur