Add YAML file for competencies and validate (#2)
This adds a YAML file for competencies and some of the tooling we need
around this. I've created a command to parse the YAML and save as JSON,
and a command to validate the generated JSON against the schema.

This also contains some rough documentation, most of which is
placeholder for now. I think it will be easier to review if we add the
language documentation etc later.