Update the CircleCI config
This now includes a deploy job which is run on a release, but only after
the website has been published. This allows deploy jobs to trigger
something that _pulls_ from the new live JSON on the website.

I've also overhauled the CI config a bit, adding in caching because this
was already quite slow. Speed of building gets more and more important
as we add multiple deploy targets. This lead to some minor changes to
the Makefile as well, to optimise the build process.

This work is required before any of us can actually start writing deploy
scripts.