Deploy competency data to Google Sheets (#29)
This adds in a deployment script to push competency data to Google
Sheets. This get the most basic possible sheet set up as a
proof-of-concept, and I anticipate us changing this a lot.

The generated sheet is viewable here:
https://docs.google.com/spreadsheets/d/1V0LIbCQtJsi2iowfJnRTDr4Na4LhNAlJ_UHl9dDQs00/edit

The deployment script runs whenever a new semver release is created on
the repository. There are a few things that need to happen to make sure
this is maintainable by others:

1. I need to create a shared LastPass folder or note for any keys
2. I need to give more people access to the Google API project