Make the UUID regex more lenient
PR #30's build failed initially because a v1 UUID was used rather than
v4. There's no issue with using older or newer UUIDs, so I've updated
the validation to allow for versions 1–5.

This probably also highlights that we should provide some guidance in
the documentation on where to get your IDs. Or possibly we should add a
command-line tool to help generate them.