Write up more full documentation
This adds much more comprehensive documentation and a full contributing
guide (apart from one TODO item). I made the decision to split the
documentation across several Markdown files in the `docs` folder. These
files are heavily inter-linked.

Reviewing this may be easier to do by viewing this branch on GitHub and
starting with the README:

https://github.com/Financial-Times/engineering-progression/tree/documentation#readme