Reword opening bit
Change wording because "one source of truth for Engineering careers and progression in the CTO organisation of the Financial Times" is not quite right. It is the single source of truth for the competencies, but we have lots of places where *progression* is talked about.

I've also tried to remove the vagueness of 
"We also intend on these career competencies being made available in other familiar formats, for example a generated Google Sheet which managers and reports can use to track progress."
by rewording.