Add placeholder text to the website (#50)
For launch, the website should probably contain more than just "example"
over and over again. I've written something that is imperfect, based on
the rest of the documentation.

The page mostly just links out to different areas of the repo, but now
it at least it can act as a landing page for the whole project.