Make competency and level data available in site (#56)
This makes competency and level data available in the Jekyll site, so we
can render a nice HTML page of competencies. The page I've added isn't
linked to anywhere, and is just there to demonstrate how to use the
competencies data.

I've also added instructions for local development, because you have to
build these files manually.

This checks some of the points in #55.