Fix the website build process (#60)
We can't only build competencies data on a version change, due to the
more complicated way the website is built. Because the generated site is
committed to the gh-pages branch rather than the actual Jekyll site, we
can't only generate the `_data` folder on a valid tag. This means that
the competencies on the website directly reflect the YAML rather than
the last _version_ of the competencies. I may be able to address this
later, but for now the build is broken.