Add charts to the spreadsheet (#67)
This adds a bar chart to each level spreadsheet, which adds quite a lot
of extra code to the Google Sheets deploy script. I've tried to abstract
where I can, but we have to create four extra hidden calculation sheets
(to store the aggregated chart data) as well as a chart for each level.