Remove the hard-coded links from link headings (#69)
o-layout was generating another link and wrapping it around the
hard-coded ones. I think it's OK to remove the hard-coded ones for now,
though it means that the heading links are only present if JavaScript
loads.

It is possible to disable the auto-linking in o-layout manually, but:

1. It's not documented (we're doing this now)
2. It doesn't have a check in place to make sure that the heading
   doesn't already contain a link. I'm going to suggest this

Once we've done these things then I may add hard-coded links back in.