Add automated accessibility tests to the website
I've just been doing this work for the Origami website, and thought it
might be nice to run some automated accessibility tests on the
engineering progression website.

This now uses Pa11y CI to test some of the more obvious accessibility
issues. The build will fail if any are introduced.