Relax the requirement that all competencies be met
This PR updates the "how-to-use" guide to indicate that not all
competencies must be met for an engineer to be considered for promotion.
This is the opposite to the way we've been responding to certain issues
and questions, and we'll be updating our responses to those once this
wording has been agreed on.

We're making this change for a couple of reasons:

  1. This is a less big jump for people from the current competencies
     matrix to the new one - it's a lot to expect particularly in the
     first trials of a new system

  2. We've had feedback on several competencies that they are not easily
     meetable by engineers in certain teams. We don't want to remove or
     relax too many competencies, and allowing them to be marked as not
     met but with an explanation helps to reduce edge-cases

  3. The messaging used by JK and the tech directors so far indicates
     that not all competencies must be met. The "how-to-use" guide was
     directly contradicting this