Remove very difficult to meet leadership competency
Issue 84

This competency is very difficult to meet in the day-to-day runnings
of a developer. The examples especially are of quite complex systems
which would require the developer frequently encountering particular
systems.

So, we're going to remove this competency and add one about bootcamps
or secondments which will help encourage engineers to gain experience
outside of their own teams.