Create a clearer thread around presenting work
I've updated the language for the set of competencies that relate to
an engineer presenting their work to people. I've removed the line:

> Clearly communicates technical concepts to non-technical colleagues

This was repeated for each competency and relates more to communication.
I'll address this separately.

I've also removed a duplicate competency from M->S1 leadership.

This work should hopefully address part of #96 and close #130.