Rework "communicates tech to non tech" competencies
I'm not 100% happy with this wording, but struggling to come up with an
alternative. I wanted to align these two competencies, and there are
other competencies that introduce a single word (like "complex") to
differentiate between mid and senior.

I've also changed "non-technical" to "less technical", as I think
communicating to someone in a role like design or UX should help meet
this competency, and it seems a little wrong to call these people
non-technical.

This work should resolve #96, alongside #138