Add a competency around privacy by design
Closes #116

We have competencies around other things that indicate our systems are
well built, things about accessibility, testing, monitoring, but nothing
about privacy. This commit adds a competency to address that area.
I've added this as an M->S1 competency because the responsibility
for this being implemented feels like a senior trait, however this isn't
to say that junior engineers should be writing systems that don't have
these traits, only that they might not necessarily be able to
articulate that that is what they are doing.