Change wording about being the go-to person
We had some really good feedback on this competency - that it encourages
people to become "key person dependencies" or that it encourages them
to "knowledge hoard"

Changing the wording on this to emphasise the knowledge sharing aspect
of being an expert. What good is an expert if they don't share their
expertise?