Add examples to feedback competency
This competency sparked a lot of discussion. Quotes from the feedback
session:

* This one may be a bit more difficult because as a junior you may not
feel comfortable giving feedback to your colleagues if they are more
senior
* In my team there are 3 seniors and a junior and I would find it very
difficult to give feedback to a more senior member of staff.
* Maybe this is a cultural thing between different teams
Is this maybe a difficult competency for somebody who's more introverted
, or working on a much more senior-heavy team?

This commit adds two examples to clarify that:
- feedback doesn't need to be direct, it can go via a line manager
- feedback doesn't need to be negative, positive feedback is really
important too
- feedback can be really small - like asking someone to add an agenda
to a meeting