Move "Considers technical direction ..." down
Feedback from the workshop with S2 engineers was that at s2 level
engineers are doing more than just considering the technical direction
but they are actively shaping it.

This commit is step one of that - to move the competency about
considering back a level.
Next point is to add a competency about shaping technical direction