Remove checkboxes from the Google Sheet
The checkboxes in the Google sheet are causing us a few issues. Mostly
they promote the idea that the competencies for a level are a checklist,
and that completing the checklist will get you promoted.

The checkboxes have now been removed from the sheet, and the chart at
the top is now based on whether any evidence has been provided for the
competencies.