Reword the how-to-use guide
Based on feedback from the last round of promotions, some people assume
that the engineering competencies is a check-list and that checking all
of the boxes guarantees you a promotion. This is not true.

While our how-to-use guide never states that you will be promoted if you
meet all of the competencies, upon re-reading we can see how you might
assume so. This PR attempts to make the competencies role in a
promotions case more clear.