Add supporting URLs to the schema
In #276 we discussed the need for competencies to have supporting URLs -
sometimes more information would be useful to help understand the
competency, and URLs allow us to do this without bloating the summary or
examples.

I decided to make supporting URLs an array of objects. Using an object
rather than just the URL as a string allows us to display different link
text to help a user understand what they're clicking on.

I've also added an empty supporting URL array to all of the
competencies. This follows our decision with competency examples: it's
useful for consumers of the API to be able to trust that certain fields
are always going to be an array, reducing logic in their code.