Remove Stalebot temporarily
So far I think we've reopened every stale issue, because often they get
marked as stale when there's a natural period of inactivity on this
project (for example when we're on hiatus, or if we miss a few weeks and
don't get around to discussing certain issues).

Issue #297 discusses some improvements to the configuration. I think we
should disable Stalebot entirely until we've made these changes,
otherwise we're likely to repeat reopening issues again.