Add guidance for delivering feedback
WHY:

As per issue 207, the requester has advised that he would like more guidance on how to deliver feedback and how to write competencies. HR have written a guide on how to deliver great feedback, using whats called the AID model which may be a useful tool to use as a template when writing up individual competencies.

WHAT:

Updated the FAQ and the How To Use framework to give individuals looking at this clear guidance to follow when writing feedback on themselves or as managers feedback on their colleagues.