Remove beta messaging from the site and docs (#338)
We need to do this before releasing v1.0.0, which I'll do once this is
merged. I decided to remove the beta banner rather than change it to
announce the stable version - this turned out to be a pain to do
(because o-message styling seems a bit broken) and I don't have the time
to properly investigate now.